+++ src/whatsapp/models/message.model.ts
@@ -29,6 +29,7 @@ export class MessageRaw {
   source_id?: string;
   source_reply_id?: string;
   chatwoot?: ChatwootMessage;
+  contextInfo?: any;
 }
 
 const messageSchema = new Schema<MessageRaw>({